+++ client/Readme.md
@@ -1,0 +1,8 @@
+# Pet App Front
+
+## Commands
+
+- npm install / npm i --force -> Instala dependencias (es importante agregar el --force, porque sino el react-google-login rompe el install)
+
+- npm start -> Empieza a correr el front
+
